+++ src/dipdup/config/__init__.py
@@ -812,10 +812,13 @@ class DipDupConfig:
 
     def set_up_logging(self) -> None:
         loglevels = {}
-        if not isinstance(self.logging, dict):
+        if isinstance(self.logging, dict):
+            loglevels = {**self.logging}
+        else:
             loglevels['dipdup'] = self.logging
             loglevels[self.package] = self.logging
 
+        # NOTE: Environment variables have higher priority
         if env.DEBUG:
             loglevels['dipdup'] = 'DEBUG'
             loglevels[self.package] = 'DEBUG'