+++ src/dipdup/indexes/evm_events/index.py
@@ -1,6 +1,5 @@
 from collections import deque
 from collections.abc import Iterable
-from typing import TYPE_CHECKING
 from typing import Any
 
 from dipdup.config.evm_events import EvmEventsHandlerConfig
@@ -22,33 +21,11 @@ from dipdup.prometheus import Metrics
 QueueItem = tuple[EvmEventData, ...] | RollbackMessage
 EvmDatasource = EvmSubsquidDatasource | EvmNodeDatasource
 
-if TYPE_CHECKING:
-    from dipdup.context import DipDupContext
-
 
 class EvmEventsIndex(
     EvmIndex[EvmEventsIndexConfig, QueueItem, EvmDatasource],
     message_type=SubsquidMessageType.logs,
 ):
-    def __init__(
-        self,
-        ctx: 'DipDupContext',
-        config: EvmEventsIndexConfig,
-        datasources: tuple[EvmDatasource, ...],
-    ) -> None:
-        super().__init__(ctx, config, datasources)
-        self._topics: dict[str, dict[str, str]] | None = None
-
-    @property
-    def topics(self) -> dict[str, dict[str, str]]:
-        if self._topics is None:
-            self._topics = {}
-            for handler_config in self._config.handlers:
-                typename = handler_config.contract.module_name
-                event_abi = self._ctx.package.get_converted_abi(typename)['events']
-                self._topics[typename] = {k: v['topic0'] for k, v in event_abi.items()}
-
-        return self._topics
 
     async def _synchronize_subsquid(self, sync_level: int) -> None:
         first_level = self.state.level + 1
@@ -107,7 +84,11 @@ class EvmEventsIndex(
         handlers: tuple[EvmEventsHandlerConfig, ...],
         level_data: Iterable[EvmEventData],
     ) -> deque[Any]:
-        return match_events(self._ctx.package, handlers, level_data, self.topics)
+        event_abis = {
+            handler.contract.module_name: self._ctx.package.get_converted_abi(handler.contract.module_name)['events']
+            for handler in handlers
+        }
+        return match_events(self._ctx.package, handlers, level_data, event_abis)
 
     async def _call_matched_handler(
         self,