+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog], and this project adheres to [Semantic Versioning].
 
+## [Unreleased]
+
+### Fixed
+
+- config: Fixed setting logging levels according to the config.
+- evm.events: Fixed matching logs when filtering by topic0.
+
 ## [7.5.6] - 2024-05-16
 
 ### Fixed